python/requirements.txt
qrcode[pil]
pynacl
click